requirements.txt
#python version: 3.10.11
pandas==1.5.3
scikit-learn==0.24.1
numpy==1.26.0
matplotlib==3.8.1
seaborn==0.13.2
xgboost==2.0.3
lightgbm==4.3.0
joblib==1.3.2
category-encoders==2.6.3
scipy==1.11.2
fpdf==1.7.2
pyyaml